Refactor RPC stuff to its own submodule
